Add address and pointer pseudo watch point logic
  When either uaddr() returns a given address or a given pointer value
is returned from an allocation routine or the pointer is released, the
backtrace is printed, per print_ubacktrace().
  This is only useful if address randomization is disabled, so that
it is possible to know beforehand the value of an address or pointer;
obviously not in all cases, and if in an special case, should rebuild
with some way to synchronize events.
  On purpose it does not accept a list of addresses or pointers, to avoid
too much overhead, and need for hash table lookups.