First version of mmap.stp
  This initial code probes mmap, mremap, munmap and brk syscall, and get
as much information as possible from userlevel.

  The output is a list of mmap chunk sizes and the backtrace from where it
was allocated.

  The idea is to attempt to help in debugging code that allocates too much
memory. As long as allocation is not spread all around it should help in
finding where in code memory allocation is being done, as the memory
management code will need to either use the brk or mmap syscalls, to either
increase or create the heap, or return the memory request.

  The code is built somewhat close to the more detailed report of memory.stp
but should help in a different use case pattern when code starts allocating
too much memory and it is not spread too much in the code.