Add some possibly redundant, but useful information to logs
  It is important to have this extra output as state is only released when
the reset function is called, so, it is important to know if a data type
has been already released. Static initializers are more complicated to
handle, and should be assumed if it does not tell the state. If it is a
global variable it is easy to figure; if it is a stack variable it might
not be trivial to debug any problem without a rebuild.