Avoid too many false positives due to already exited threads
  There is no proper logic to handle synchronous thread exit, so, a map
of uncertain_threads is kept, in case these threads did just exit the
thread entry routine. This is basically managed in a probe when the
kernel finishes reaping the tid. The logic is a bit complicated because
it also does not have any kind of "synchronous" way to detect the process
is exiting, and reaping too early could print misleading unexpected state
messages, while it is really desired to know what was the state when the
process was terminated.