Add initial version of the glibc.stp script
  This script attempts to debug glibc and only glibc internal allocations.
  There might be some false positives, that can be debugged with:

-Gdebug_called_from_glibc_p=1

and silenced with:

-Gdebug_called_from_glibc_p=1
-Gignore_mismatch_in_called_from_glibc_p=1

  These are disabled by default as they might require too many map entries
to record all allocations and detect the mismatch. Mismatches might happen
based on stack frame layouts. An initial, commented heuristic to detect a
few cases has been commented, to avoid hiding bugs. Better stack frame parsing
could/should be added to avoid as much as possible these debug options.

  There is also the option:

-Grecord_double_free=1

to debug double free. It is there for convenience as it is very unlikely
to have a double free in a released glibc.